+++ src/MeasurementSets.jl
@@ -75,6 +75,7 @@ include("tables/create.jl")
 include("tables/edit.jl")
 include("tables/resync.jl")
 include("taql/commands.jl")
+include("beam/beam.jl")
 
 export AbstractTable, Table, RefTable, ConcatTable
 export readtable, columnnames, columndesc, keywords, subtables, nrow
@@ -89,6 +90,10 @@ export MEpoch, MDirection, MPosition, MFrequency, MRadialVelocity, RefFrame, ref
 export MBaseline, MuvW, MEarthMagnetic, IGRF, earthfield
 export EarthMagneticMachine, emm_lineofsight
 export rotation_measure, faraday_rotation, derotate_angle, RM_IONOSPHERE
+export PrimaryBeam, GaussianBeam, AiryBeam, PolynomialBeam
+export EllipticalGaussianBeam, SquintBeam
+export power_response, voltage_response, attenuate, correct_flux, angular_separation, reffreq,
+    pointing_offset
 export Ephemeris, open_ephemeris, field_ephemeris, ephemeris_direction,
     ephemeris_radvel, ephemeris_distance, ephemeris_diskpos
 export UTC, TAI, TT, TDB, UT1, J2000, ICRS, B1950, APP, GALACTIC, ECLIPTIC,