+++ test/runtests.jl
@@ -43,6 +43,7 @@ const SAMPLE_MS = get(ENV, "MEASUREMENTSETS_TEST_MS",
     include("taql_query_tests.jl")
     include("taql_command_tests.jl")
     include("units_tests.jl")
+    include("beam_tests.jl")
     if isdir(SAMPLE_MS)
         include("measures_tests.jl")
         include("taql_mscal_tests.jl")